% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myconstr.R
\name{myconstr}
\alias{myconstr}
\title{My constructor}
\usage{
myconstr(x, y, alpha)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}

\item{alpha}{level of significance}
}
\value{
a named list of class Rttest containing a data
       frame containing x and y, alpha, the confidence
       interval for mux - muy and the pvalue
}
\description{
a constructor function makes an object of a particular OOP
class and makes a lot of useful statistical information that
methods operate on to produce whatever particulars we need
}
\examples{
\dontrun{set.seed(21);x <-rnorm(30,5,2); set.seed(23); y<- rnorm(40, 3,2);alpha <- 0.05
obj <- myconstr(x = x, y = y, alpha = 0.05)}
}
