% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{Plot method for class Rttest}
\usage{
\method{plot}{Rttest}(x, ...)
}
\arguments{
\item{x}{object made from my constr}

\item{...}{plot variables}
}
\value{
a side by side boxplot of the data
}
\description{
Plot method for class Rttest
}
\examples{
\dontrun{plot(obj)}
}
